Test: extract some code to utils
