+++ tests_config/utils/consistent-end-of-line.js
@@ -1,0 +1,13 @@
+"use strict";
+
+function consistentEndOfLine(text) {
+  let firstEndOfLine;
+  return text.replace(/\r\n?|\n/g, (endOfLine) => {
+    if (!firstEndOfLine) {
+      firstEndOfLine = endOfLine;
+    }
+    return firstEndOfLine;
+  });
+}
+
+module.exports = consistentEndOfLine;
